LagartoDOMBuilderTagVisitor shouldn't log errors when error logging is disabled
Motivation:

`errorLogEnabled` config option was introduced to disable error logging.
But `LagartoDOMBuilderTagVisitor#error` keeps on calling the ErrorLogConsumer even when errorLogEnabled is set to false.

Users can configure a noop ErrorLogConsumer instead of the default one, but still this is a workaround, not the desired behavior.

Modification:

Only call ErrorLogConsumer when error logging is enabled.

Result:

errorLogEnabled works as expected: when set to false, no error should be logged.